+++ R/kinfitr_bloodmodels.R
@@ -725,7 +725,9 @@ blmod_exp_startpars <- function(time, activity, fit_exp3=TRUE,
 #' @param multstart_upper Optional. The upper limits of the starting parameters.
 #' @param multstart_iter The number of fits to perform with different starting
 #'   parameters. If set to 1, then the starting parameters will be used for a
-#'   single fit.
+#'   single fit. Otherwise a single number selects the starting values by
+#'   improved Latin hypercube sampling over the multstart bounds rather than at
+#'   random, while one value per parameter forms a Cartesian grid.
 #' @param Method_weights If no weights provided, should the weights be divided
 #'   by discrete and continuous samples equally (i.e. with more continuous
 #'   samples, the continuous samples each get less weight). Default is TRUE.
@@ -910,7 +912,7 @@ blmod_exp <- function(time, activity, Method = NULL,
 
     multmodelout_multi <- nls.multstart::nls_multstart(
       formula = as.formula(formula), modelweights = weights,
-      data = blood, iter = multstart_iter,
+      data = blood, iter = multstart_iter, lhstype = "improved",
       start_lower = multstart_lower, start_upper = multstart_upper,
       supp_errors = "Y", lower = lower, upper = upper)
 
@@ -1629,7 +1631,9 @@ blmod_feng_startpars <- function(time, activity,
 #' @param multstart_upper Optional. The upper limits of the starting parameters.
 #' @param multstart_iter The number of fits to perform with different starting
 #'   parameters. If set to 1, then the starting parameters will be used for a
-#'   single fit.
+#'   single fit. Otherwise a single number selects the starting values by
+#'   improved Latin hypercube sampling over the multstart bounds rather than at
+#'   random, while one value per parameter forms a Cartesian grid.
 #' @param Method_weights If no weights provided, should the weights be divided
 #'   by discrete and continuous samples equally (i.e. with more continuous
 #'   samples, the continuous samples each get less weight). Default is TRUE.
@@ -1791,7 +1795,7 @@ blmod_feng <- function(time, activity, Method = NULL,
 
     multmodelout_multi <- nls.multstart::nls_multstart(
       formula = as.formula(formula), modelweights = weights,
-      data = blood, iter = multstart_iter,
+      data = blood, iter = multstart_iter, lhstype = "improved",
       start_lower = multstart_lower, start_upper = multstart_upper,
       supp_errors = "Y", lower = lower, upper = upper)
 
@@ -1961,7 +1965,9 @@ predict_blood_feng <- function(object, newdata = NULL) {
 #' @param multstart_upper Optional. The upper limits of the starting parameters.
 #' @param multstart_iter The number of fits to perform with different starting
 #'   parameters. If set to 1, then the starting parameters will be used for a
-#'   single fit.
+#'   single fit. Otherwise a single number selects the starting values by
+#'   improved Latin hypercube sampling over the multstart bounds rather than at
+#'   random, while one value per parameter forms a Cartesian grid.
 #' @param Method_weights If no weights provided, should the weights be divided
 #'   by discrete and continuous samples equally (i.e. with more continuous
 #'   samples, the continuous samples each get less weight). Default is TRUE.
@@ -2162,7 +2168,7 @@ blmod_fengconv <- function(time, activity, inftime = NULL,
 
     multmodelout_multi <- nls.multstart::nls_multstart(
       formula = as.formula(formula), modelweights = weights,
-      data = blood, iter = multstart_iter,
+      data = blood, iter = multstart_iter, lhstype = "improved",
       start_lower = multstart_lower, start_upper = multstart_upper,
       supp_errors = "Y",
       lower = lower, upper = upper,
@@ -2376,7 +2382,9 @@ predict_blood_fengconv <- function(object, newdata = NULL) {
 #' @param multstart_upper Optional. The upper limits of the starting parameters.
 #' @param multstart_iter The number of fits to perform with different starting
 #'   parameters. If set to 1, then the starting parameters will be used for a
-#'   single fit.
+#'   single fit. Otherwise a single number selects the starting values by
+#'   improved Latin hypercube sampling over the multstart bounds rather than at
+#'   random, while one value per parameter forms a Cartesian grid.
 #' @param Method_weights If no weights provided, should the weights be divided
 #'   by discrete and continuous samples equally (i.e. with more continuous
 #'   samples, the continuous samples each get less weight). Default is TRUE.
@@ -2568,7 +2576,7 @@ blmod_fengconvplus <- function(time, activity, inftime = NULL,
 
     multmodelout_multi <- nls.multstart::nls_multstart(
       formula = as.formula(formula), modelweights = weights,
-      data = blood, iter = multstart_iter,
+      data = blood, iter = multstart_iter, lhstype = "improved",
       start_lower = multstart_lower, start_upper = multstart_upper,
       supp_errors = "Y", lower = lower, upper = upper,
       convergence_count = FALSE,